requirements.txt
Mako==1.1.4
MarkupSafe==2.0.1
PyYAML==5.4.1
